Partial views for add areas and add systems sucessfully created

+++ Controllers/CreateTemplatesController.cs
@@ -12,33 +12,59 @@ namespace HHI_InspectionSoftware.Controllers
     {
         private HHIEntities6 db = new HHIEntities6();
 
-        // GET: CreateTemplates
-        public ActionResult Index()
-        {
-            return View();
-        }
+        //// GET: CreateTemplates
+        //public ActionResult Index()
+        //{
+        //    return View();
+        //}
 
         // GET: CreateTemplates/Create
-        public ActionResult Create()
+        public ActionResult Save()
         {
             return View("Save");
         }
 
-        public ActionResult SaveGet(int? templateID, TemplateModel templateModel)
-        {
-            //TemplateModel templateModel = new TemplateModel();
-            //templateModel.Template = db.Templates.Find(templateID);
-            //templateModel.Areas = areas;
-
-            return View("Save", templateModel);
-        }
-
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Save(TemplateModel templateModel)
+        public ActionResult Update(TemplateModel templateModel)
         {
             if (ModelState.IsValid)
             {
+                // Checks to see if template name is available
+                // Aborts if not, saves if so
+                bool nameTaken = false;
+                foreach(var t in db.Templates)
+                {
+                    if(templateModel.Template.Name == t.Name)
+                    {
+                        nameTaken = true;
+                    }
+                }
+                if (nameTaken)
+                {
+                    // name taken - abort save - find out how to prompt user to change name through view 
+                    return View("Save", templateModel);
+                }
+                else
+                {
+                    db.Templates.Add(templateModel.Template);
+                    db.SaveChanges();
+                }
+            }
+            return View("Save", templateModel);
+        }
+
+        //public ActionResult SaveGet(int? templateID, TemplateModel templateModel)
+        //{
+        //    return View("Save", templateModel);
+        //}
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult SaveArea(TemplateModel templateModel)
+        {
+            //if (ModelState.IsValid)
+            //{
                 var name = templateModel.Template.Name;
                 bool doesExist = false;
                 int templateID = 0;
@@ -62,8 +88,81 @@ namespace HHI_InspectionSoftware.Controllers
                     templateModel.Area.TemplateID = templateID;
                     db.Areas.Add(templateModel.Area);
                     db.SaveChanges();
-
+                    ViewBag.AreaID = templateModel.Area.ID;
                 }
+
+            //}
+            foreach (var a in db.Areas)
+            {
+                if (a.TemplateID == templateID)
+                {
+                    templateModel.Template.Areas.Add(a);
+                }
+            }
+
+            foreach (var s in db.HomeSystems)
+            {
+                if (s.TemplateID == templateID)
+                {
+                    templateModel.Template.HomeSystems.Add(s);
+                }
+            }
+
+            templateModel.Template.ID = templateID;
+
+
+            if(templateModel.Template.Areas != null)
+            {
+                ViewBag.Areas = new SelectList(templateModel.Template.Areas, "ID", "Name");
+            }
+            else
+            {
+                ViewBag.Areas = new SelectList("");
+            }
+            if (templateModel.Template.HomeSystems != null)
+            {
+                ViewBag.Systems = new SelectList(templateModel.Template.HomeSystems, "ID", "Name");
+            }
+            else
+            {
+                ViewBag.Systems = new SelectList("");
+            }
+
+            return View("Save", templateModel);
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult SaveSystem(TemplateModel templateModel)
+        {
+            //if (ModelState.IsValid)
+            //{
+                var name = templateModel.Template.Name;
+                bool doesExist = false;
+                int templateID = 0;
+                foreach (var t in db.Templates)
+                {
+                    if (t.Name == name)
+                    {
+                        doesExist = true;
+                        templateID = t.ID;
+                    }
+                }
+                if (!doesExist)
+                {
+                    db.Templates.Add(templateModel.Template);
+                    db.SaveChanges();
+                    templateID = templateModel.Template.ID;
+                }
+
+                if (templateModel.HomeSystem.Name != null)
+                {
+                    templateModel.HomeSystem.TemplateID = templateID;
+                    db.HomeSystems.Add(templateModel.HomeSystem);
+                    db.SaveChanges();
+                    ViewBag.SystemID = templateModel.HomeSystem.ID;
+                }
+
                 foreach (var a in db.Areas)
                 {
                     if (a.TemplateID == templateID)
@@ -71,13 +170,7 @@ namespace HHI_InspectionSoftware.Controllers
                         templateModel.Template.Areas.Add(a);
                     }
                 }
-                if (templateModel.HomeSystem.Name != null)
-                {
-                    templateModel.HomeSystem.TemplateID = templateID;
-                    db.HomeSystems.Add(templateModel.HomeSystem);
-                    db.SaveChanges();
 
-                }
                 foreach (var s in db.HomeSystems)
                 {
                     if (s.TemplateID == templateID)
@@ -85,25 +178,52 @@ namespace HHI_InspectionSoftware.Controllers
                         templateModel.Template.HomeSystems.Add(s);
                     }
                 }
-                return View(templateModel);
-               
+                templateModel.Template.ID = templateID;
+
+                //ViewBag.Areas = new SelectList(templateModel.Template.Areas, "ID", "Name");
+                //ViewBag.Systems = new SelectList(templateModel.Template.HomeSystems, "ID", "Name");
+
+                //return View("Save", templateModel);
+            //}
+
+            if (templateModel.Template.Areas != null)
+            {
+                ViewBag.Areas = new SelectList(templateModel.Template.Areas, "ID", "Name");
             }
-            
-            return View();
+            else
+            {
+                ViewBag.Areas = new SelectList("");
+            }
+            if (templateModel.Template.HomeSystems != null)
+            {
+                ViewBag.Systems = new SelectList(templateModel.Template.HomeSystems, "ID", "Name");
+            }
+            else
+            {
+                ViewBag.Systems = new SelectList("");
+            }
+
+            return View("Save", templateModel);
         }
 
-        public ActionResult DeleteArea(int? id, int? templateID)
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult DeleteArea(TemplateModel templateModel)
         {
             //int templateID = templateModel.Template.ID;
-            Area area = db.Areas.Find(id);
-            db.Areas.Remove(area);
-            db.SaveChanges();
-            TemplateModel templateModel = new TemplateModel();
-            templateModel.Template = db.Templates.Find(templateID);
+            //TemplateModel templateModel = new TemplateModel();
             List<Area> areas = new List<Area>();
+            Area area = db.Areas.Find(templateModel.Area.ID);
+            if(area != null)
+            {
+                db.Areas.Remove(area);
+                db.SaveChanges();
+               // templateModel.Template = db.Templates.Find(templateID);
+            }
+
             foreach (var a in db.Areas)
             {
-                if (a.TemplateID == templateID)
+                if (a.TemplateID == templateModel.Template.ID)
                 {
                     areas.Add(a);
                 }
@@ -115,12 +235,17 @@ namespace HHI_InspectionSoftware.Controllers
         public ActionResult DeleteSystem(int? id, int? templateID)
         {
             //int templateID = templateModel.Template.ID;
-            HomeSystem system = db.HomeSystems.Find(id);
-            db.HomeSystems.Remove(system);
-            db.SaveChanges();
             TemplateModel templateModel = new TemplateModel();
-            templateModel.Template = db.Templates.Find(templateID);
             List<HomeSystem> systems = new List<HomeSystem>();
+            HomeSystem system = db.HomeSystems.Find(id);
+
+            if(system != null)
+            {
+                db.HomeSystems.Remove(system);
+                db.SaveChanges();
+                templateModel.Template = db.Templates.Find(templateID);
+            }
+
             foreach (var s in db.HomeSystems)
             {
                 if (s.TemplateID == templateID)
@@ -132,23 +257,53 @@ namespace HHI_InspectionSoftware.Controllers
             return View("Save", templateModel);
         }
 
-        public ActionResult CreateTemplate([Bind(Include = "ID,Name")] Template template, List<Area> areas)
+        public ActionResult SaveLimitation(TemplateModel templateModel)
         {
             if (ModelState.IsValid)
             {
-                db.Templates.Add(template);
+                Limitation limitation = templateModel.Limitation;
+                db.Limitations.Add(limitation);
                 db.SaveChanges();
-                int templateID = db.Templates.Max(item => item.ID);
-
-                foreach (var a in areas)
+            }
+            List<Limitation> limitations = new List<Limitation>();
+            foreach(var l in db.Limitations)
+            {
+                if(l.AreaID == templateModel.Area.ID)
                 {
-                    a.TemplateID = templateID;
-                    db.Areas.Add(a);
-                    db.SaveChanges();
+                    limitations.Add(l);
                 }
             }
+            templateModel.Limitations = limitations;
 
-            return View();
+            return View("Save", templateModel);
         }
+
+        public void SaveCheckItem(CheckItem checkItem)
+        {
+            if (ModelState.IsValid)
+            {
+                db.CheckItems.Add(checkItem);
+                db.SaveChanges();
+            }
+        }
+
+        //public ActionResult CreateTemplate([Bind(Include = "ID,Name")] Template template, List<Area> areas)
+        //{
+        //    if (ModelState.IsValid)
+        //    {
+        //        db.Templates.Add(template);
+        //        db.SaveChanges();
+        //        int templateID = db.Templates.Max(item => item.ID);
+
+        //        foreach (var a in areas)
+        //        {
+        //            a.TemplateID = templateID;
+        //            db.Areas.Add(a);
+        //            db.SaveChanges();
+        //        }
+        //    }
+
+        //    return View();
+        //}
     }
 }
